+++ src/components/LandingPage/Vitrine/Vitrininhas/Vitrine.js
@@ -1,16 +1,16 @@
 import React, { Component } from 'react'
 import { useState } from "react";
 import styled from 'styled-components'
-import { anunciosAulas } from './Anuncio'
+import { anunciosAulas } from '../Anuncio'
 import ArrowForwardIosIcon from "@material-ui/icons/ArrowForwardIos";
 import ArrowBackIosIcon from "@material-ui/icons/ArrowBackIos";
 
 
 
 const ContainerVitrine = styled.div`
- width: 19%;
+  width: 19%;
   height: 30vh;
-  color:#000072;
+  color:white;
 
   .Expositor {
     height: 100%;
@@ -44,7 +44,7 @@ const ContainerVitrine = styled.div`
     background-color: rgb(0, 0, 0, 0.0);
     display: grid;
     place-items: center;
-    color:#b7b7c3;
+    color: #b7b7c3;
     cursor: pointer;
   }
   .Centro {
@@ -70,19 +70,20 @@ const ContainerVitrine = styled.div`
 
   h6,
   p {
-    background-color: rgb(183, 183, 195, 0.8);
+    background-color: rgb(172, 23, 104,0.8);
     padding: 3px;
     border-radius: 2px;
+    
   }
 `
 
-function VitrineAulas() {
-  const [currImg, setCurrImg] = useState(0);
+function Vitrine() {
+  const [currImg, setCurrImg] = useState(2);
 
   return (
     <ContainerVitrine>
       <a>
-      <div  className="Expositor"  style={{ backgroundImage: `url(${anunciosAulas[2].img})` }}>
+      <div  className="Expositor"  style={{ backgroundImage: `url(${anunciosAulas[3].img})` }}>
         
         <div className="DivBotao1"
         onClick={() => {
@@ -93,7 +94,7 @@ function VitrineAulas() {
 
         <div className="Centro">
           <h6>{anunciosAulas[currImg].title}</h6>
-          <p>{anunciosAulas[currImg].price} <br/> {anunciosAulas[currImg].contact}</p>
+          <p>{anunciosAulas[currImg].price} |<br/> {anunciosAulas[currImg].contact}</p>
           </div>
 
         <div className="DivBotao2" 
@@ -109,4 +110,4 @@ function VitrineAulas() {
     </ContainerVitrine>
   );
 }
-export default VitrineAulas;
+export default Vitrine;
