+++ src/components/LandingPage/Vitrine/Vitrininhas/VitrineOutros.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { useState } from "react";
 import styled from 'styled-components'
-import { anunciosOutros } from './Anuncio'
+import { anunciosOutros } from '../Anuncio'
 import ArrowForwardIosIcon from "@material-ui/icons/ArrowForwardIos";
 import ArrowBackIosIcon from "@material-ui/icons/ArrowBackIos";
